Correction du répertoire en ne tenant pas compte des .git

+++ gitlab-emothep.py
@@ -130,10 +130,14 @@ class GitEmothepGitlab(object):
         for dir in os.listdir():
             src_path = configfile.LOCALREPO+"/"+configfile.PATH_REPO_PACKAGES
             for package in os.listdir():
-                print('Delete symlink in packages repository')
-                dst_path = configfile.SAGHOME+"/"+package
-                os.unlink(dst_path)
-                shutil.move(src_path, dst_path)
+                if not package.startswith('.git'):
+                    print('Delete symlink in packages repository')
+                    dst_path = configfile.SAGHOME+"/"+package
+                    print('dst_path : %s'% dst_path)
+                    os.unlink(dst_path)
+                    print('Symnlink removed')
+                    shutil.move(src_path, dst_path)
+                    print('Package moved')
 
     def __removeLink(self,projectName):
         print('Revert symlink to package : %s'% projectName)
@@ -142,13 +146,14 @@ class GitEmothepGitlab(object):
             src_path = configfile.LOCALREPO+"/"+configfile.PATH_REPO_PACKAGES
             print('src_path : %s'% src_path)
             for package in os.listdir():
-                print('Delete symlink in packages repository')
-                dst_path = configfile.SAGHOME+"/"+package
-                print('dst_path : %s'% dst_path)
-                os.unlink(dst_path)
-                print('Symnlink removed')
-                shutil.move(src_path, dst_path)
-                print('Package moved')
+                if not package.startswith('.git'):
+                    print('Delete symlink in packages repository')
+                    dst_path = configfile.SAGHOME+"/"+package
+                    print('dst_path : %s'% dst_path)
+                    os.unlink(dst_path)
+                    print('Symnlink removed')
+                    shutil.move(src_path, dst_path)
+                    print('Package moved')
 
     
     def __statusProject(self, projectName, all) :
